ci: make the pre-commit hooks pass on the pull-request runner
Every pull-request run since August failed the "checks" job on the base branch
itself. The hooks now pass locally on the whole tree:

- check-shebang-scripts-are-executable: the six benchmark scripts carry a
  shebang; mark them executable in the index.
- ruff check: two en dashes in fdtdmex.io.mock docstrings (RUF002), an
  unused unpack, two redundant int(round()) casts and a list concatenation in
  the validation tests (RUF059, RUF046, RUF005).
- ruff format: five files the formatter rewraps.
- ty: scope the checker to the library. examples/, benchmarks/ and server/
  are scripts, notebooks and the MCP server and are excluded like tests/;
  unresolved-import is ignored because mlx only installs on Apple silicon and
  the runner is Linux (pyvista and plotly are optional extras for the same
  reason). The library's own diagnostics are fixed in the next commit.

+++ tests/validation/test_subpixel.py
@@ -77,7 +77,7 @@ def test_smoothing_reduces_mode_staircase_error():
     k0 = 2 * np.pi / lam
     n_core, n_clad = 1.5, 1.0
     Lx, dx, fac = 6e-6, 40e-9, 20
-    nx = int(round(Lx / dx))
+    nx = round(Lx / dx)
     edges = (np.arange(nx + 1) - nx / 2) * dx
     centers = 0.5 * (edges[:-1] + edges[1:])
     fcent = (np.arange(nx * fac) - nx * fac / 2 + 0.5) * (dx / fac)
